requirements.txt
langchain_community==0.3.21
neo4j==5.28.1
langchain_openai==0.3.12
# Fixed typo: "langchian" -> "langchain". Use a compatible 0.3.x constraint.
langchain>=0.3.0,<0.4.0
langgraph==0.3.31
langgraph-prebuilt==0.1.8
python-dotenv>=1.0.0

# HuggingFace embeddings (free alternative to OpenAI)
langchain-huggingface>=0.1.0
sentence-transformers>=2.2.0

# Google Gemini (Removed)
# google-generativeai>=0.3.0

fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
requests>=2.28.0
pytest>=7.0.0
transformers>=4.30.0
torch>=2.0.0
httpx>=0.24.0
